bot/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

aiocache == 0.11.1
fastapi == 0.86.0
pydantic == 1.10.2
python_i18n == 0.3.9
python_telegram_bot == 13.14
requests == 2.28.1
uvicorn == 0.18.2
